refactor: organize runway map layering
